@@ -248,11 +248,11 @@ export class GeradorCobol {
     let dots = "";
     const missingDotsNumber = Colunas.COLUNA_FIM - lineText.length;
     if (missingDotsNumber > 0) {
-      for (var i = 1; i <= missingDotsNumber; i++) {
+      for (let i = 1; i <= missingDotsNumber; i++) {
         dots = dots.concat(".");
       }
       await this.editor.setCurrentLine(lineText + dots);
     }
   }
 
-};
+}
